requirements.txt
django
djangorestframework
djangorestframework-simplejwt
social-auth-core
social-auth-app-django
